requirements.txt
pyrebase
python-dotenv
flask
argon2-cffi
gunicorn
libgravatar
requests
re
